refactor(ds): extract ConversationInspector to features/chat (5d.5)
Final per-feature inspector split. Moves the conversation render
block (~509 lines) + all conversation/intake editor state + 5
handlers out of InspectorPanel into a focused feature-local component.

ConversationInspector (727 lines)
---------------------------------
- src/features/chat/components/ConversationInspector.tsx
- Consumes useUserDetail + useMatterDetail + usePracticeDetail
  concurrently (conversation shows client + linked matter + practice
  context simultaneously)
- Owns three sub-paths driven by mode + isClientView flag:
  1. PRACTICE_ONBOARDING → delegates to SetupInspectorContent
  2. isClientView → branded InspectorHeaderHero + intake status +
     consultation detail editor (intake fields: name/email/phone +
     practiceServiceUuid + city + state + opposingParty +
     desiredOutcome + hasDocuments + householdSize + courtDate +
     description)
  3. regular → InspectorHeaderPerson + contacts row + linked-matter +
     assignee + priority + tags editors
- Owns state: activeConversationEditor (14-position discriminator with
  intake sub-editors), localIntakeDraft, skipBlurRef, isSavingAssignment/
  Priority/Tags/Matter, localError
- Owns memos: priorityOptions, assignedToOptions, currentTags, tagOptions,
  matterOptions, intakeServiceOptions, assignedMemberLabel,
  assignedConversationMember, currentMatterLabel, currentPriorityLabel,
  currentTagsLabel, conversationPeople
- Owns handlers: handleConversationAssignmentChange / PriorityChange /
  TagsChange / MatterChange, handleIntakeFieldChange
- Renders its own loading skeleton + error banner
- Display token cleanups: text-input-placeholder → text-dim,
  text-input-text → text-ink, text-red-400 → text-neg

InspectorPanel dispatcher cleanup
---------------------------------
InspectorPanel is now a true thin dispatcher (243 lines, down from
1004 lines pre-5d.5 and 2005 lines at session start):
- Imports + delegates: <ConversationInspector .../> with all conv
  props pass-through when entityType === 'conversation'
- Removed massive amount of conversation-specific code:
  - 9 useState slots + skipBlurRef
  - 13 memos (priorityOptions, assignedToOptions, currentTags,
    tagOptions, matterOptions, intakeServiceOptions, assignedMemberLabel,
    assignedConversationMember, currentMatterLabel, currentPriorityLabel,
    currentTagsLabel, conversationPeople, currentAssignedLabel)
  - 5 handlers (assignment / priority / tags / matter / intake)
  - All 3 data hook calls (per-feature inspectors own their own data)
  - Reset useEffects (each per-feature inspector resets its own)
  - resolveString helper
- Removed unused imports (massive list — see InspectorPanel diff)
- Net InspectorPanel change: -797 lines deleted, +34 lines added

Trajectory: 2005L (start) → 1925L (post-5c.4 invoice) → 1241L
(post-5d.3 client) → 1004L (post-5d.4 matter) → 243L (post-5d.5
conversation). All per-feature inspector content now lives in
features/*; InspectorPanel is just a switch + header chrome.

5d.6 (delete dispatcher + 4 callers dispatch directly) follows in the
next commit.
